initialize: better TUI while installing

+++ initialise.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
+error() {
+	echo -e "\e[1;31m$1\e[0m"
+}
+
+alreadyDone() {
+	echo -e "\e[1;32m$1\e[0m"
+}
+
 # Download and install latest updates.
 sudo apt update
 sudo apt upgrade -y
@@ -8,7 +16,7 @@ sudo apt upgrade -y
 sudo ubuntu-drivers autoinstall
 
 # This will no longer be needed when I setup debugger in nvim.
-sudo snap install code --classic
+# sudo snap install code --classic
 
 ################
 #  ESSENTAILS  #
@@ -39,49 +47,86 @@ echo 'export PAGER=most' >> ~/.bashrc
 #############
 #  LAZYGIT  #
 #############
-LAZYGIT_VERSION=$(curl -s "https://api.github.com/repos/jesseduffield/lazygit/releases/latest" | grep -Po '"tag_name": "v\K[0-9.]+')
-MACHINE_NAME=$(uname --machine)
-curl -Lo lazygit.tar.gz "https://github.com/jesseduffield/lazygit/releases/latest/download/lazygit_${LAZYGIT_VERSION}_Linux_${MACHINE_NAME}.tar.gz"
-sudo tar xf lazygit.tar.gz -C /usr/local/bin lazygit
-rm -rf lazygit.tar.gz
+if [[ $(which lazygit) != 0 ]]; then
+	error "Installing lazygit"
+	LAZYGIT_VERSION=$(curl -s "https://api.github.com/repos/jesseduffield/lazygit/releases/latest" | grep -Po '"tag_name": "v\K[0-9.]+')
+	MACHINE_NAME=$(uname --machine)
+	curl -Lo lazygit.tar.gz "https://github.com/jesseduffield/lazygit/releases/latest/download/lazygit_${LAZYGIT_VERSION}_Linux_${MACHINE_NAME}.tar.gz"
+	sudo tar xf lazygit.tar.gz -C /usr/local/bin lazygit
+	rm -rf lazygit.tar.gz
+else
+	alreadyDone "lazygit is already insalled with verison $(lazygit --version)"
+fi
 
 ###############
 #  XOURNAL++  #
 ###############
-sudo add-apt-repository ppa:apandada1/xournalpp-stable
-sudo apt update
-sudo apt install xournalpp
+if [[ $(which xournalpp) != 0 ]]; then
+	error "Installing xournalpp"
+	sudo add-apt-repository ppa:apandada1/xournalpp-stable
+	sudo apt update
+	sudo apt install xournalpp -y
+else
+	alreadyDone "xournalpp is already insalled with verison $(lazygit --version)"
+fi
 
 ############
 #  NEOVIM  #
 ############
-wget https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.deb
-sudo apt install ./nvim-linux64.deb
+
+if [[ $(which nvim) != 0 ]]; then
+	error "Installing nvim"
+	wget https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.deb
+	sudo apt install ./nvim-linux64.deb
+else
+	alreadyDone "Neovim is already installed with version $(nvim --version)"
+fi
 
 	###########################
 	#  NODEJS - for nvim lsp  #
 	###########################
-	curl -sL https://deb.nodesource.com/setup_16.x | sudo bash -
-	sudo apt install nodejs
+	if [[ $(which node) != 0 ]]; then
+		error "Installing nodejs"
+		curl -sL https://deb.nodesource.com/setup_16.x | sudo bash -
+		sudo apt install nodejs -y
+	else
+		alreadyDone "Nodejs is already insalled with version $(node --verison)"
+	fi
 
 	#####################
 	#  RUST - for nvim  #
 	#####################
-	curl https://sh.rustup.rs -sSf | sh
+	if [[ $(which cargo) != 0 ]]; then
+		error "Installing Rust (cargo)"
+		curl https://sh.rustup.rs -sSf | sh
+	else
+		alreadyDone "Rust is already intalled with version $(cargo --version)"
+	fi
 
 ###########
 #  BRAVE  #
 ###########
-sudo curl -fsSLo /usr/share/keyrings/brave-browser-archive-keyring.gpg https://brave-browser-apt-release.s3.brave.com/brave-browser-archive-keyring.gpg
-echo "deb [signed-by=/usr/share/keyrings/brave-browser-archive-keyring.gpg arch=amd64] https://brave-browser-apt-release.s3.brave.com/ stable main"|sudo tee /etc/apt/sources.list.d/brave-browser-release.list
-sudo apt install brave-browser
+
+if [[ $(which brave-browser) != 0 ]]; then
+	error "Installing brave"
+	sudo curl -fsSLo /usr/share/keyrings/brave-browser-archive-keyring.gpg https://brave-browser-apt-release.s3.brave.com/brave-browser-archive-keyring.gpg
+	echo "deb [signed-by=/usr/share/keyrings/brave-browser-archive-keyring.gpg arch=amd64] https://brave-browser-apt-release.s3.brave.com/ stable main"|sudo tee /etc/apt/sources.list.d/brave-browser-release.list
+	sudo apt install brave-browser -y
+else
+	alreadyDone "Brave is already installed with version $(brave-browser --version)"
+fi
 
 #############
 #  SPOTIFY  #
 #############
-curl -sS https://download.spotify.com/debian/pubkey_5E3C45D7B312C643.gpg | sudo apt-key add - 
-echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
-sudo apt install spotify-client
+if [[ $(which spotify) != 0 ]]; then
+	error "Installing spotify"
+	curl -sS https://download.spotify.com/debian/pubkey_5E3C45D7B312C643.gpg | sudo apt-key add -
+	echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
+	sudo apt install spotify-client -y
+else
+	alreadyDone "Spotify is already installed with verison $(spotify --version)"
+fi
 
 ###########
 #  GNOME  #
